feat: passing proof
